Bluetooth: controller: Add support for ISO feature bits
Add support for ISO feature bits. This includes extending fields that
hold features to 64 bits to be able to handle the host-controlled
"Isochronous Channels" feature.

+++ subsys/bluetooth/controller/ll_sw/ull_conn_types.h
@@ -162,8 +162,8 @@ struct ll_conn {
 	struct {
 		uint8_t  req;
 		uint8_t  ack;
-		uint32_t features_conn;
-		uint32_t features_peer;
+		uint64_t features_conn;
+		uint64_t features_peer;
 	} llcp_feature;
 
 	struct {
